refactor: LoginMember 패키지 변경
