+++ src/main/java/roomescape/auth/controller/AuthController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.auth.dto.LoginMember;
 import roomescape.auth.service.AuthService;
 import roomescape.client.CookieManager;
 import roomescape.member.dto.request.LoginRequest;
