+++ src/main/java/roomescape/auth/dto/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.auth.controller;
+package roomescape.auth.dto;
 
 import roomescape.member.domain.Role;
